Add files via upload
Everything from my lowduino project that I archived from my old website. 

An Arduino Nano runs a custom piece of code averages readings from the fuel sender and then references a fuel tank 'map' to work out the corresponding PWM value to send to the fuel gauge to show the correct level. The averaging prevents the gauge needle from oscillating back and forth as fuel moves in the tank as the vehicle is in motion. 

Use the fuel mapping spreadsheet to calculate the map for a given tank and replace the values in the code.

Parts list is in the BOM PDF along with what was then (back in 2010) Farnell part numbers. I believe the .brd, .sch & .pde files are from Eagle... I can not recall for certain which ED package I used, but I think Eagle may be the one.